chore: move subscription deferral info to value objects
